pyproject.toml
[tool.poetry]
name = "tripwire"
version = "0.1.0"
description = "A macOS portscan honeypot"
authors = ["Your Name <you@example.com>"]
license = "GNU GPLv3"

[tool.poetry.dependencies]
python = "^3.9"
scapy = "^2.4.5"
ntfy = "^2.7.0"
pystreams = "^0.6"
PyYAML = "^5.4.1"
validators = "^0.18.2"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
